+++ util-core/src/main/scala/com/twitter/io/exp/MinimumThroughput.scala
@@ -129,6 +129,8 @@ object MinimumThroughput {
         }
       }
     }
+
+    def close(deadline: Time): Future[Unit] = writer.close(deadline)
   }
 
   /**